Refactors as lib/gambit
