+++ lib/gambit.js
